+++ src/types.rs
@@ -4,16 +4,16 @@
 pub struct NumNodes(pub u64);
 
 #[derive(PartialEq, Eq, Copy, Clone, Show)]
-pub struct NumVariations(pub u16);
+pub struct NumVariations(pub u32);
 
 #[derive(PartialEq, Eq, Copy, Clone, Show)]
-pub struct PerMill(pub u16);
+pub struct PerMill(pub u32);
 
 #[derive(PartialEq, Eq, Copy, Clone, Show)]
-pub struct NumCpu(pub u16);
+pub struct NumCpu(pub u32);
 
 #[derive(PartialEq, Eq, Copy, Clone, Show)]
-pub struct NumPlies(pub u16);
+pub struct NumPlies(pub u32);
 
 #[derive(PartialEq, Eq, Copy, Clone, Show)]
-pub struct NumMoves(pub u16);
+pub struct NumMoves(pub u32);
